requirements.txt
smartapi-python==1.3.4
websocket-client==1.6.4
pyotp==2.9.0
logzero==1.7.0
python-dotenv==1.0.0
twilio==8.10.0
APScheduler==3.10.4
SQLAlchemy==2.0.36
pandas>=3.0.2
requests==2.31.0
pytz==2024.1
